Add default backend for load API
