+++ python/paddle/jit/dy2static/program_translator.py
@@ -20,7 +20,6 @@ import os
 import threading
 import warnings
 import weakref
-from enum import Enum, auto
 from typing import TYPE_CHECKING, Any, Callable, Generic, TypeVar
 
 from typing_extensions import ParamSpec, Self
@@ -88,29 +87,6 @@ __all__ = []
 MAX_TRACED_PROGRAM_COUNT = 10
 
 CONVERSION_OPTIONS = "__jst_not_to_static"
-
-
-class Backend(Enum):
-    CINN = auto()
-    PHI = auto()
-
-    @staticmethod
-    def from_arg(arg: str | Backend | None):
-        if isinstance(arg, Backend):
-            return arg
-        if arg is None:
-            return Backend.PHI
-        if arg.upper() == "CINN":
-            return Backend.CINN
-        raise ValueError(
-            f"Unknown backend {arg}. Only support 'CINN' or None for PHI."
-        )
-
-    def is_cinn(self):
-        return self == Backend.CINN
-
-    def is_phi(self):
-        return self == Backend.PHI
 
 
 def synchronized(func):
